+++ app/models/contact.rb
@@ -7,10 +7,11 @@ class Contact < ApplicationRecord
   validates :email, format: {with: URI::MailTo::EMAIL_REGEXP}, allow_blank: true
 
   enum category: {
-    other: 0,
-    school: 1,
-    trust: 2,
+    diocese: 4,
     local_authority: 3,
-    diocese: 4
+    school: 1,
+    solicitor: 5,
+    trust: 2,
+    other: 0
   }
 end